image_to_vm.sh: Fix some bugs in the debug output code around the stateful dd.
There were two bugs in this code (at least). The first bug was that the
stateful partition was chmod-ed only the first time around, and after blockdev
--rereadpt was run the partition file was apparently recreated. We need to run
chmod every time around in order to have it work.

Second, the expression which decided whether to run blockdev vs. just trying
again was not correct as far as bash syntax goes. I tested this new version in
a small test script, and it has the right behavior.

BUG=chromium:418650
TEST=Inspection and running portions of the script in a specialized test
script.

Change-Id: I7e5720815e85c297e7dba0e70ac9e7259598092b
Reviewed-on: https://chromium-review.googlesource.com/221040
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>