loopback: Expand the sanity checking of loopback partition device nodes.
It turns out that not only will the kernel not always find the partitions in
when setting up a loopback device, it will also not always set them up
correctly. Gwendal correctly predicted that the start offset of the partitions
would be wrong in those cases, so we can check for that as well.

BUG=chromium:418650
TEST=Ran image_to_vm.sh (the consumer of loopback devices) manually and
verified that it completed successfully. Ran it with artificial errors
injected into the checks and saw that they were reported and that the dev was
rescanned like it was supposed to be.

Change-Id: I4f8957b6765d890131daaf109027f940102eadf5
Reviewed-on: https://chromium-review.googlesource.com/221564
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>