image_to_vm: Remove hack to rename USB partition.
image_to_vm used to rename the USB rootfs partition from /dev/sdb to /dev/sda.
This was required because the rootfs for USB boot was hard coded to /dev/sdb,
and in the absence of any other drive, this had to be renamed to /dev/sda for VM
image.

After CL:225840, we use PARTUUID to boot from USB, so this hack is no longer
needed.

BUG=chromium:425210
TEST=Successfully boot a VM image for link.

Change-Id: Ic29ba300c1c54e459fa962908fbb361c207159d2
Reviewed-on: https://chromium-review.googlesource.com/227162
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>