build_image/image_to_vm: improve output/error checking
The current build_image finished with a general statement for the command
to run.  This example though only works for dev images, so trying to use
it with any other type fails.  Rework the output so it reflects each type
the user has actually generated.

On the image_to_vm side, the error checking of user arguments is a bit
thin.  It tries to run even when the inputs don't exist leading to a lot
of confusing output.  Clean that up and issue clear messages when a path
that should exist does not.

BUG=chromium:434674
TEST=`./build_image ... base dev test` summarizes things correctly for each image type

Change-Id: I0c657882ea69df2be7b088a939a06e19ed64b5f2
Reviewed-on: https://chromium-review.googlesource.com/231960
Reviewed-by: Stefan Sauer <ensonic@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>