build_packages: install debug symbols by default
Install debug symbols at the end of build_packages if they are available. This
is turned on by default. It will be off by default once the mechanism is rolled
out to every board and is stable.

This is a copy of https://chromium-review.googlesource.com/#/c/233301/

BUG=chromium:426494
TEST=trybot run on gizmo, link, daisy, lumpy-incremental.

CQ-DEPEND=CL:234857

Change-Id: I913238d8743b2a0fef97f4a265a1e4d7d237c3bf
Reviewed-on: https://chromium-review.googlesource.com/234829
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>