disk_layout.json: Describe reserved erase block counts
This patch makes disk_layout.json include a count of reserved erase
blocks for bad block mapping. This is necessary to include so that
- The reserved space can be included in the partition size used in
  cgpt add (not yet implemented) on top of the size
- The reserved erase blocks can be passed in the kernel command line
  to ubi, which will then check that they are in fact reserved
- Our kernel command line building code checks that enough blocks
  are reserved to make it very unlikely that we will run out.

UBI by default expects 2% of the device to be reserved in each MTD
partition. For us, that would mean 10-12% of the device reserved.
This handles the absolute worst case, but in fact on small partitions
you don't need to reserve quite so much. You do have to reserve
proportionately more in small partitions, however. Overall, for the
first board that I updated this way (the storm-nand variant), 2.8%
of the device is reserved for the stateful and rootfs partitions,
vs 6% that would be reserved otherwise.

This patch changes cgpt.py to pass through the reserved block count
as well as to check that the values used are reasonable in its
'validate' functionality (which replaces parseonly). It's useful
to put the checking somewhere in our code because when we pass the
reserved block count to UBI, that overrides UBI's built-in checking.
Developers will be able to edit disk_layout.json with more confidence
on a new board knowing that the reserved_block_count is tested to
be reasonable--they will just have to update the max_bad_block_count
in accordance with the NAND datasheet.

In disk_layout.json, a new 'metadata' partition is added as a
possibility. This creates a place to put key/value information on
a per-target basis. Previously, key/value pairs would have to be in
the board-wide metadata, or pertain to a particular partition. Now,
a "num": "metadata" partition can have key/value information about
a particular target for a board. This metadata section is used for
the new NAND parameters, as well as for the "hybrid_mbr" option.

BUG=chromium:442776
TEST=Tried various values for reserved_blocks and saw reasonable-
looking probabilities for failures, or successes.
TEST=Observed the reserved block count passed to the kernel in
the kernel command line.
TEST=Omitted the device size, used a small reserve, and observed
the checks to be bypassed and a warning to be printed
TEST=Tested hybrid MBR option still works. This patch is is a
CQ-loop with hybrid MBR update patches
CQ-DEPEND=CL:*192885

Change-Id: I50c85d66397e595613386bfa8795fbfc3dae01b4
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236611
Reviewed-by: Mike Frysinger <vapier@chromium.org>