zero_free_space: Don't fill rootfs with zero when not mounted.
The zero_free_space function creates a file with zeros to fill all
the rootfs' data blocks on a ext2 filesystem. When the filesystem is
copied from the read-only filesystem to directory on the the host
filesystem, this function would fill the host filesystem instead
of the rootfs.

This patch skips this step when the rootfs directory isn't a
mountpoint.

BUG=chromium:382951
TEST=`./build_image --board=link` works and shows the log message of
'Zeroing freespace'

Change-Id: I2667f7a4e885816e24886224f5d944a0da6aaf51
Reviewed-on: https://chromium-review.googlesource.com/240392
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>