_dump_trace: stop accessing out of bounds $BASH_SOURCE[$#FUNCNAME[@]]
In some conditions hard to reproduce, bash stops being lenient and
stops tolerating access to the last+1 element of the BASH_SOURCE[]
array. When that happens it throws a "BASH_SOURCE[n]: unbound
variable" error. Combined with errexit (set -e), this error stops the
execution a split second before printing the trace, which proved to be
very time-consuming.

This bug has gone unnoticed so far because bash is usually lenient and
returns an (unused) empty string. I noticed it because I've been
intensively re-using the _dump_trace function for temporary tracing and
"reverse-engineering".

BUG=None (yet!)
TEST=run "./build_image foobar" before and after change and see no change.

Change-Id: I79acb729e431626f95830954d4907d14a14cc22e
Reviewed-on: https://chromium-review.googlesource.com/241381
Commit-Queue: Marc Herbert <marc.herbert@intel.com>
Tested-by: Marc Herbert <marc.herbert@intel.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>