build_kernel_image: Iterate over image types to find NAND params
When adding parameters to the kernel command line to initialize UBI
and ubiblock, build_kernel_image.sh is invoked for the usb image_type
but the NAND partitions show up elsewhere, likely in base. As a result,
./build_image was not adding NAND parameters. This patch makes
build_kernel_image.sh iterate over all image types and add NAND
parameters from any of them that needs it. An alternative would be to just
add NAND parameters for base, but that seems like a hack to me. Support is
added in cgpt.py to expose the list of image types.

TEST=Observed the right kernel cmdline params to be emitted on storm_nand
BRANCH=none
BUG=chromium:427951

Change-Id: I56b40b94833ccf3fcbc2f937c29c1fd041a313cc
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241530
Reviewed-by: Mike Frysinger <vapier@chromium.org>