New mount/unmount wrappers with squashfs rw support.
squashfs doesn't support to be mounted as read-write as it doesn't have
a data structure to handle modifications efficiently. Instead, you need
to rebuild the entire filesystem whenever you want to modify it.

This patch introduces two wrappers "fs_umount" and "fs_mount" that
support handling a squashfs in rw mode by copying or "unsquashing" the
contents of the filesystem to the mountpoint directory and recreating
the squashed filesystem and writing it back to the partition.

The initial creation of the filesystem done by mk_fs now handles a
squashfs fs_format.

BUG=chromium:444311
TEST=`./build_image --board=cosmos` using a fs_format=squashfs in the
layout creates a squashsfs rootfs (and fails to mount it later).
TEST=`./build_image` still works in normal builds.

Change-Id: I3abdb8d2640e0f9c7da82a87ec1921dfa2e60b2c
Reviewed-on: https://chromium-review.googlesource.com/247334
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>