cgpt.py: Add check for total device size
To make it easier to develop disk_layout.json for small, fixed-size
boards, this patch adds a check to see if all partitions exceed the
size of the whole storage device. This check is only performed if
the device size is included in the layout's metadata.

TEST=For storm_nand, played around with increasing partition sizes and
reserved block counts and observed the size check to fail due to those
when appropriate.
BUG=chromium:442776

Change-Id: I50cb29b06be0848ffff1aa36af6c8d3a550be84f
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243318
Reviewed-by: Mike Frysinger <vapier@chromium.org>