nand_disk_layout.json: Reserve space for BBT in NAND
The MTD bad block table (BBT) code doesn't need to be told where to
store the bad block table. However, it'll reserve a block for itself
towards the end of the device. This patch makes nand_disk_layout.json
include a BBT partition with four blocks for the BBT. It also makes a
minor update to cgpt.py so that it doesn't choke on an empty partition
with reserved erase blocks.

BUG=none
TEST=builds Storm image

Change-Id: Id42d0232aaa77fe045a916910eefcfcde6d6d234
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243383