mod_image_for_recovery: Make the EFI partition optional
On boards with raw NAND, to save space and because legacy EFI boot will not
work, I left out the EFI partition. This patch makes mod_image_for_recovery.sh
still work in the case where the disk layout does not include an EFI partition.

BUG=none
TEST=Built a recovery image for storm_nand, which seems to be validated by
vboot but hangs in the transition from firmware to kernel

Change-Id: Iaae79a501d7e5748595da768fe876a12df2674dc
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/244141