cgpt.py: Fix handling of empty partition when stacking layouts.
cgpt.py allows to stack layouts to split the configuration. If the
layout defines an empty partition (without a partition number) in
order to leave a gap and it inherits from another layout where the
gap wasn't present, the new gap will be ignored.

In order to fix this, we need to ensure that the order of the
partitions in the parent layout and the new layout is the same, so
the gaps (partitions without number) are placed in the same order.

This patch handles the empty partitions properly in this case, so
far unused in Chromium OS, and implements unittests for cgpt.py.

CQ-DEPEND=CL:245170,CL:*194852,CL:*194879,CL:*194879
BUG=chromium:454046
TEST=build_library/cgpt_unittest.py
TEST=for f in ../*overlays/*/scripts/disk_layout.json; do
  build_library/cgpt.py validate base "$f" || echo "$f failed";
done

Change-Id: I97bd857f871247cbb4f434654a34b57dc63beccf
Reviewed-on: https://chromium-review.googlesource.com/245163
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>