cgpt.py: Include fs_options in the partition script.
For squashfs filesystem, the options used to create the filesystem
are required if you want to modify it, in order to unsquash and
squash it again.

This patch includes those options when generating the partition
script.

BUG=chromium:444311
TEST=`build_library/cgpt.py write base .../disk_layout.json output.sh`
includes the FS_OPTIONS_*.

Change-Id: I73f81745612fe4028feff3effe12ea4e07bf7c72
Reviewed-on: https://chromium-review.googlesource.com/245653
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>