mount_gpt_image.sh: Use fs_mount/fs_umount to support squashfs in rw.
When mounting an image as read-write, if the image contains a squashfs
partition, we need to unsquash and resquash the filesystem. For this
purpose we use the fs_mount/fs_umount wrappers instead of mounting the
partition directly.

This patch requires to pass the same image parameters to the --umount
operation that were used when mounting. In the cases where this is not
passed, we emit a warning and fallback to just umount all the
filesystems directly.

BUG=chromium:455814,brillo:161
TEST=`./build_image --board=cosmos` with a rootfs in squashfs
mounts/unmounts the chromiumos_image.bin.
TEST=`./build_image --board=urara`

Change-Id: Ibeebeda2885edeb58ddf89e8d24283914e00eb50
Reviewed-on: https://chromium-review.googlesource.com/247336
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>