cros_make_image_booteable: Remove --enable_squashfs support.
cros_make_image_booteable supports replacing the ext2 rootfs partition
with a squashed version of it. This was introduced in 2011, but the
approach of post-processing the rootfs to convert it to a squashfs
doesn't work if the original data doesn't fit in the available space
for the rootfs partition formated as ext2. If you are using squashfs,
presumably you need it to store more data than it fits uncompressed
in the partition, so this approach will fail.

While we develop a different approach to support squashfs, integrated
in mount_gpt_image, we want to remove this conflicting feature.

BUG=chromium:382951
TEST=cbuildbot amd64-generic-full arm-generic-full

Change-Id: I4f3b937da5770f7cbebd190202e58a2217449abb
Reviewed-on: https://chromium-review.googlesource.com/246860
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>