cgpt.py: Account for UBI metadata in partition sizes
UBI partitions take two pages per block and 4 blocks per
partition in metadata. This patch reserves that amount both
when checking if the partitions sum up to too large a size
and when actually sizing the partitions on flash with cgpt.

BUG=none
TEST=Built for urara and found there to be exactly enough
space for the rootfs, with the 'size' listed to be equal for
NAND and USB in disk_layout.json
CQ-DEPEND=CL:*198095
CQ-DEPEND=CL:*198125

Change-Id: I547f64ef257aa023d50f77e6492ed4725e52c22c
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/247741
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Nam Nguyen <namnguyen@chromium.org>