cgpt.py: Allow a blank EFI partition
For NAND boards, we have a blank EFI partition since legacy boot
won't work anyway. Previously, whether or not an EFI partition
was present, a 'cgpt boot' call was made to make the partition
bootable. This patch makes that call conditional on whether the
partition is blank.

Nevertheless, cgpt boot needs to be called to make a proper USB
image. So this patch adds a non-blank EFI partition to the USB
image for NAND boards.

BUG=none
TEST=called cgpt.py write on cosmos and storm overlays; observed
cgpt boot in storm and not in cosmos, as appropriate
Successfully built and booted from a USB image on a NAND device.

Change-Id: Iea4f01c8fd9aa33541da8815d90b68bf1d7adaf4
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/247740
Reviewed-by: Mike Frysinger <vapier@chromium.org>