chromeos-common.sh: Remove link to real chromeos-common.sh.
chromeos-common.sh really exists in
src/platform2/installer/share/chromeos-common.sh. A symlink to that
file was left in src/scripts, but unused outside of
generate_au_zip.py (already addressed).

This symlink breaks in minilayout and SDK builds. To ensure we aren't
using it, and that it doesn't break anything in the future, remove it.

BUG=brillo:364
TEST=Local/remote builds.

Change-Id: Ide6f2304ff2c2c58c43a57d0ea7b96738a34ec48
Reviewed-on: https://chromium-review.googlesource.com/253145
Reviewed-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>