_process_mount_pt: always set up target when force is enabled
If the dir we try to set up here already exists, we end up skipping the
setup logic for the source symlink.  This makes sense normally (e.g. when
we enter an existing chroot), but when we're doing an initial bootstrap,
we want to always set it up regardless of the source paths.  We already
have a force flag for this, but the force wasn't forceful enough -- it
only bypassed the inner logic.  Make it so it bypasses the initial test
as well.

BUG=chromium:460815
TEST=chromiumos-sdk passes

Change-Id: I5d58c7a3126c752841291af7cd21defbdda537c7
Reviewed-on: https://chromium-review.googlesource.com/253551
Trybot-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>