build_kernel_image.sh: Handle squashfs rootfs.
build_kernel_image needs to read the size of the rootfs (in 4KiB
blocks) that is then used to create the verity hash tree. This patch
extends this logic to support squashfs when computing the size of the
filesystem.

BUG=chromium:382951
TEST=./build_image --board=cosmos dev test # with cosmos changes to use squashfs

Change-Id: I1e5393e70e18ae7908476cacd8db4b2e2193a078
Reviewed-on: https://chromium-review.googlesource.com/255601
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>