setup_board: partial support for brick
This CL provide partial support for bricks to setup_board.
We can build a functional sysroot for a given brick using the --brick
argument. The toolchain support is not complete yet: the toolchain can
be detected for a given brick but we can't install the toolchain if it
is missing.

BUG=brillo:502
TEST=`setup_board --brick //foo` works. Packages can be built with
emerge and parallel_emerge.

CQ-DEPEND=CL:259000
CQ-DEPEND=CL:258770
CQ-DEPEND=CL:258522

Change-Id: I8164f99d13bec32d0461ddcc2bfd9516e2fbf570
Reviewed-on: https://chromium-review.googlesource.com/258810
Reviewed-by: Steve Fung <stevefung@chromium.org>
Trybot-Ready: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Bertrand Simonnet <bsimonnet@chromium.org>