mount_gpt_util.sh: Allow --safe and --read_only on mount_image.
remount_image lets you re-mount in read-only or safe mode (rootfs is
read-only and the rest is read-write) but you can't mount the image
read-only or safe directly. In a normal build_image step for base
images, you don't need to mount the image read-write when running
cros_make_image_bootable so that step can be avoided, improving
performance in the squashfs case as mounting the image read-write
is expensive. The image will be mounted as read-write when building
dev and test images since those use --force_developer_mode.

This patch allows the --safe and --read_only flags to be passed to
mount_image and uses it when possible in cros_make_image_bootable.

This patch also includes a check on {disable|enable}_rw_mount that the
passed block device is read-write. This is required because mount
will create a read-only loop device when mounting a filesystem
read-only, as opposed to mounting it read-write and then remounting
it read-only.

BUG=None
TEST=`./build_image --board=link dev test`.

Change-Id: I1c20d69d91c653a0e7d48d46afd0bff7d4395186
Reviewed-on: https://chromium-review.googlesource.com/257930
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>