setup_board: Clear the board value if brick was provided.
In a previous change we overrode the board value using the brick's
friendly name, but this caused other logic to fail because the two
values were not expected. Instead, we simply clear the board value. This
reinstates the invariant that exactly one of them must be set.

BUG=None
TEST=None

Change-Id: I6159cfcb8097bd8afbd88c1cce72a2b1d2900061
Reviewed-on: https://chromium-review.googlesource.com/262480
Trybot-Ready: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>