build_image: Report an error when the squashfs is too big.
This patch adds a check that the sizelimit used to mount/umount the
filesystem is respected for the squashfs filesystem. This is important
since the filesystem is assembled at the end of the installation, so
even if you installed too many packages, the problem doesn't manifest
until the end of the installation. Currently, when the filesystem
doesn't fit in the specified fs_size, but fits in the partition size,
the verity hashes will destroy part of the filesystem and you wouldn't
notice until after the image booted.

BUG=chromium:463793
TEST=Added and manually ran unittest. ./build_image fails when fs_size
is too small.

Change-Id: Ic21a97a76d2f3a28bf64238735e6ba05c585fa81
Reviewed-on: https://chromium-review.googlesource.com/262937
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>