update hooks: force java update
When we update java in the sdk, the old version stays installed and the
active jvm is that.  We need to make sure the new one is installed and
we switch to the new jvm.  Otherwise, people with existing installs will
stay on the old one while new sdks roll out with the new one.

BUG=chromium:346369
TEST=`cbuildbot panther_moblab-release` passes
CQ-DEPEND=CL:272058

Change-Id: Ia5e8350d260724b1b0bb6951cbce29e830e3f709
Reviewed-on: https://chromium-review.googlesource.com/273961
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>