legacy_disk_layout: Increase STATE size by 20%.
The statefulpartition (STATE) is expanded to the rest of the device
size after installing or running recovery, which is normally at least
10 GiB. The image file built in the host has a fixed size big enough to
hold all the files installed in the stateful partition.

Normal (base) images don't install much on the stateful partition, but
dev and test images install all the extra dev and test dependencies in
the stateful partition. Recent builds have reached the 1 GiB limit of
test dependencies.

This patch increases the size of the stateful partition in all images
by 20%.

BUG=chromium:516978
TEST=`./build_library/cgpt.py debug 2gb-rootfs build_library/legacy_disk_layout.json`
shows the correct size (1216 MiB) for STATE. This implies that parsing
and format are correct.

Change-Id: Ia97a6a71b98cd810f716cd9269a5312b4e84c723
Reviewed-on: https://chromium-review.googlesource.com/291257
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>