Clean portage cache.
libchromeos-ui was moved out of libchromeos. As we install libchromeos'
shared libraries with libchromeos-*.so, any old objects with that name
will get installed. In particular, any previously built libchromeos-ui
library which will then conflict with chromeos-base/libchromeos-ui.
Note: this is only a problem when libchromeos is cros worked on.

BUG=None
TEST=trybot.

Change-Id: Ia4c162d5e6cddda5a1656372d2fcb7a70db78c76
Reviewed-on: https://chromium-review.googlesource.com/296695
Commit-Ready: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>