scripts: Usb passthrough support.
This CL adds a new argument to cros_start_vm:
  - usb_devices

This variable specifies a list of usb bus and device ids
for usb passthrough. The kvm is launched with usb passthrough
if this argument is given and each element of the list is
a colon separated pair.

BUG=chromium:532553
TEST=Launched a vm with usb passthrough and verified the device
was discovered by lsusb.

Change-Id: Ie702ada420b8d427afd1da9c31ee9bac8e0ffdca
Reviewed-on: https://chromium-review.googlesource.com/300153
Commit-Ready: Matthew Sartori <msartori@chromium.org>
Tested-by: Matthew Sartori <msartori@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>