cros_start_vm: log serial output to a file
Currently all serial output goes to /dev/null even when the image is
built with the serial console enabled.  Instead, send it to a named
file so that people can `tail` it to see what's going on.  If there
is no serial support in the image, then it'll simply be an empty log.

BUG=None
TEST=created an image with serial enabled; launched it; saw output in the log

Change-Id: I3c6a056efa6f2399ffb9aa1ce4e844370607b8a1
Reviewed-on: https://chromium-review.googlesource.com/301670
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>