build_image_util: clarify failure message
If build_image fails, it will output a list of the largest files in the
destination rootfs disk. The message is confusing because it does not
make clear that these are files in the *destination* directory rather
than the source, and that they made be partially files extracted if the
untar died in the middle of a file. Clarify this.

BUG=None
TEST=None

Change-Id: Ic4eea84983b98ad056ea9cd22be707617177ed7a
Reviewed-on: https://chromium-review.googlesource.com/333789
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>