cros_vm_lib: allow running VM from inside the chroot
The QEMU binary inside the chroot is perfectly able to run the Chrome OS
VM (with KVM support enabled).
Enable this feature and tweak the command-line accordingly if needed
as we need to use the 'qemu-system-x86_64' binary inside the chroot
with the '-enable-kvm' flag.

note: this allows to run the testing on a modern (v2.5) and consistent
version of QEMU.

BUG=chromium:605616 chromium:591626
TEST=inside the chroot './bin/cros_start_vm --board=amd64-generic'
and run 'test_that --board=amd64-generic localhost:9222 suite:smoke'
TEST='cbuildbot --remote -g 339839 amd64-generic-full'

Change-Id: I4aa317400d566ddf9518c2c1ccbe1167681ba7e7
Reviewed-on: https://chromium-review.googlesource.com/339839
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>